Separate sorting functions for different uses
The same NameSortFun was being used to sort different things, which did
not always have the same shape. Fix so separate functions are used for
the different cases.

Also, update the unit tests to cover the case where a single cluster is
configured for realtime replication with two or more other clusters.