Fix race condition in riak_snmp_stat_poller
There is a race in riak_snmp_stat_poller, which occurs when realtime
replication is enabled for more than two clusters. If the replication is
enabled quickly enough, then the bug does not manifest itself. This
commit fixes the race and also some additional bugs in a code path that
was clearly never tested.